// src/types/modules/navigations.ts
export enum ROUTES_ENUM {
	Dashboard = 'wp-swift-pwa-dashboard',
	Cache = 'wp-swift-pwa-cache',
	Manifest = 'wp-swift-pwa-manifest',
	ServiceWorker = 'wp-swift-pwa-service-worker',
	Notifications = 'wp-swift-pwa-notifications',
	Analytics = 'wp-swift-pwa-analytics',
	License = 'wp-swift-pwa-license',
}